# swagger.yaml
openapi: 3.0.0
info:
  title: Urban Farming Platform API
  description: API for Interactive Urban Farming Platform
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Development Server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          type: object
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        price:
          type: number
        category:
          type: string
        availableQuantity:
          type: integer
        isOrganic:
          type: boolean
...
    Order:
      type: object
      properties:
        id:
          type: string
        totalAmount:
          type: number
        status:
          type: string
        orderDate:
          type: string
...
    Plant:
      type: object
      properties:
        id:
          type: string
        plantName:
          type: string
        healthStatus:
          type: string
paths:
...
  /produce:
    get:
      tags:
        - Products
      summary: Get all products
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: category
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      pages:
                        type: integer

    post:
      tags:
        - Products
      summary: Create product (Vendor only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - price
                - category
                - availableQuantity
              properties:
                name:
                  type: string
                description:
                  type: string
                price:
                  type: number
                category:
                  type: string
                availableQuantity:
                  type: integer
                unit:
                  type: string
                isOrganic:
                  type: boolean
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /produce/{id}:
    get:
      tags:
        - Products
      summary: Get single product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Product'
        '404':
          description: Product not found

  # ==================== RENTALS ====================
...
  /orders:
    get:
      tags:
        - Orders
      summary: Get my orders
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Orders retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'

    post:
      tags:
        - Orders
      summary: Create order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - items
                - shippingAddress
              properties:
                items:
                  type: array
                  items:
                    type: object
                    required:
                      - produceId
                      - quantity
                    properties:
                      produceId:
                        type: string
                      quantity:
                        type: integer
                shippingAddress:
                  type: string
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # ==================== COMMUNITY ====================
...
  /plants:
    get:
      tags:
        - Plants
      summary: Get my plants
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Plants retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Plant'

    post:
      tags:
        - Plants
      summary: Add plant
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - plantName
                - plantType
                - plantingDate
              properties:
                plantName:
                  type: string
                plantType:
                  type: string
                plantingDate:
                  type: string
                  format: date
                expectedHarvestDate:
                  type: string
                  format: date
                healthStatus:
                  type: string
                notes:
                  type: string
      responses:
        '201':
          description: Plant added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...